Changed the structure of the application
