+++ src/main/scala/io/scalac/wtf/domain/WtfApplication.scala
@@ -1,0 +1,12 @@
+package io.scalac.wtf.domain
+
+import akka.actor.ActorSystem
+import akka.stream.ActorMaterializer
+
+trait WtfApplication extends App
+  with DatabaseDependancy
+  with UserRoutes {
+
+  implicit val actorSystem       = ActorSystem("wtf-system")
+  implicit val actorMaterializer = ActorMaterializer()
+}
